==> notebooks/data_engineering/03_cnh_gold_layer_aggregation_and_modeling.ipynb <==
import pandas as pd
import os



# =======================
# 0. Definir ruta base del proyecto
# =======================
ruta_base = os.path.abspath(os.path.join(os.getcwd(), ".."))
ruta_silver = os.path.join(ruta_base, "SILVER")
ruta_gold = os.path.join(ruta_base, "GOLD")
os.makedirs(ruta_gold, exist_ok=True)


# =======================
# 1. Cargar archivos Parquet desde carpeta SILVER
# =======================
df_aceite = pd.read_parquet(os.path.join(ruta_silver, "aceite_condensado.parquet"))
df_gas = pd.read_parquet(os.path.join(ruta_silver, "gas_natural.parquet"))

# Definir el nombre de la columna de producción
columna_produccion = 'petroleo_mbd'

# Agrupar por 'campo', sumar la producción, ordenar y obtener el top 10
top_10_productores = df_aceite.groupby('campo')[columna_produccion].sum().sort_values(ascending=False).head(10)

print("\n--- Top 10 Campos por Producción Total de Petróleo ---")
print(top_10_productores)

# --- 3. Crear el nuevo DataFrame con la serie de tiempo del Top 10 ---

# Obtener la lista de nombres de los campos del top 10
lista_top_10_campos = top_10_productores.index.tolist()

# Filtrar el DataFrame original para mantener solo los campos del top 10
df_top_10 = df_aceite[df_aceite['campo'].isin(lista_top_10_campos)].copy()

 # --- 4. Preparar y guardar el DataFrame final ---
# Asegurarse de que la columna de fecha esté en formato datetime
df_top_10['fecha'] = pd.to_datetime(df_top_10['fecha'])

# Ordenar los datos para un análisis de series de tiempo correcto
df_top_10.sort_values(by=['campo', 'fecha'], inplace=True)

 # Guardar el resultado en un nuevo archivo CSV
ruta_archivo_salida = "top_10_campos_produccion_timeseries.csv"
df_top_10.to_csv(ruta_archivo_salida, index=False)
    

df_top_10.to_parquet(os.path.join(ruta_gold, "top_10_campos_timeseries.parquet"), index=False)

print("Vista previa aceite:")
display(df_aceite.head(2))


print("Vista previa gas:")
display(df_gas.head(2))


df_aceite.describe()

# Ruta donde quieres guardar el archivo Excel
ruta_excel = os.path.join(ruta_gold, "aceite_limpio.xlsx")

# Exportar a Excel
df_aceite.to_excel(ruta_excel, index=False)

print(f"✅ Archivo exportado correctamente a: {ruta_excel}")

# ===============================
#  Operadores únicos - Aceite
# ===============================
operadores_aceite = sorted(df_aceite["operador"].unique())
df_operadores_aceite = pd.DataFrame(operadores_aceite, columns=["operador"])
print(f"Total de operadores únicos en aceite: {len(operadores_aceite)}")
display(df_operadores_aceite)


# =======================
# . Agregar columna de año
# =======================
df_aceite["anio"] = pd.to_datetime(df_aceite["fecha"]).dt.year
df_gas["anio"] = pd.to_datetime(df_gas["fecha"]).dt.year


# =======================
# 3. Producción total por año
# =======================
df_total_anual_petroleo = df_aceite.groupby("anio")["petroleo_mbd"].sum().round(2).reset_index()
display(df_total_anual_petroleo.head())
df_total_anual_petroleo.to_parquet(os.path.join(ruta_gold, "total_anual_petroleo.parquet"), index=False)

df_total_anual_gas = df_gas.groupby("anio")["gas_natural_sin_nitrogeno_mmpcd"].sum().round(2).reset_index()
display(df_total_anual_gas.head())
df_total_anual_gas.to_parquet(os.path.join(ruta_gold, "total_anual_gas.parquet"), index=False)


# =======================
# 4. Promedio por operador
# =======================
df_promedio_operador_petroleo = df_aceite.groupby("operador")["petroleo_mbd"].mean().round(2).reset_index()
display(df_promedio_operador_petroleo.sort_values(by="petroleo_mbd", ascending=False))
df_promedio_operador_petroleo.to_parquet(os.path.join(ruta_gold, "promedio_operador_petroleo.parquet"), index=False)

df_promedio_operador_gas = df_gas.groupby("operador")["gas_natural_sin_nitrogeno_mmpcd"].mean().round(2).reset_index()
display(df_promedio_operador_gas.sort_values(by="gas_natural_sin_nitrogeno_mmpcd", ascending=False))
df_promedio_operador_gas.to_parquet(os.path.join(ruta_gold, "promedio_operador_gas.parquet"), index=False)


df_promedio_operador_petroleo[df_promedio_operador_petroleo["operador"].str.contains("PEMEX", case=False, na=False)]


df_promedio_operador_petroleo.sort_values(by="petroleo_mbd", ascending=False).head(10)


# =======================
# 5. Top 10 campos productores
# =======================
df_top10_campos_petroleo = df_aceite.groupby(["campo", "cuenca", "operador"])["petroleo_mbd"].sum().round(2).reset_index().sort_values(by="petroleo_mbd", ascending=False).head(10)
display(df_top10_campos_petroleo)
df_top10_campos_petroleo.to_parquet(os.path.join(ruta_gold, "top10_campos_petroleo.parquet"), index=False)

df_top10_campos_gas = df_gas.groupby(["campo", "cuenca", "operador"])["gas_natural_sin_nitrogeno_mmpcd"].sum().round(2).reset_index().sort_values(by="gas_natural_sin_nitrogeno_mmpcd", ascending=False).head(10)
display(df_top10_campos_gas.head())
df_top10_campos_gas.to_parquet(os.path.join(ruta_gold, "top10_campos_gas.parquet"), index=False)

# =======================
# 6. Operadores distintos
# =======================
df_operadores_distintos_petroleo = df_aceite[["operador"]].drop_duplicates()
display(df_operadores_distintos_petroleo)
df_operadores_distintos_petroleo.to_parquet(os.path.join(ruta_gold, "operadores_distintos_petroleo.parquet"), index=False)

df_operadores_distintos_gas = df_gas[["operador"]].drop_duplicates()
display(df_operadores_distintos_gas)
df_operadores_distintos_gas.to_parquet(os.path.join(ruta_gold, "operadores_distintos_gas.parquet"), index=False)


# =======================
# 7. Operador top 1 anual por volumen de gas
# =======================
df_operador_anual_gas = df_gas.groupby(["anio", "operador"])["gas_natural_sin_nitrogeno_mmpcd"].sum().reset_index()
df_top1_operador_anual_gas = df_operador_anual_gas.sort_values(["anio", "gas_natural_sin_nitrogeno_mmpcd"], ascending=[True, False]).groupby("anio").head(1)
display(df_top1_operador_anual_gas)
df_top1_operador_anual_gas.to_parquet(os.path.join(ruta_gold, "top1_operador_gas_anual.parquet"), index=False)



# =======================
# 8. Operador top 1 anual por volumen de petróleo
# =======================
df_operador_anual_petroleo = df_aceite.groupby(["anio", "operador"])["petroleo_mbd"].sum().reset_index()
df_top1_operador_anual_petroleo = df_operador_anual_petroleo.sort_values(["anio", "petroleo_mbd"], ascending=[True, False]).groupby("anio").head(1)
display(df_top1_operador_anual_petroleo)
df_top1_operador_anual_petroleo.to_parquet(os.path.join(ruta_gold, "top1_operador_petroleo_anual.parquet"), index=False)


# Top 10 operadores por promedio de producción de petróleo
top10_operadores_aceite = (
    df_aceite.groupby("operador")["petroleo_mbd"]
    .mean()
    .round(2)
    .sort_values(ascending=False)
    .head(10)
    .reset_index()
)

display(top10_operadores_aceite)



# Guardar en Parquet
top10_operadores_aceite.to_parquet(
    os.path.join(ruta_gold, "top10_operadores_petroleo.parquet"), index=False
)

# Top 10 operadores por promedio de producción de gas
top10_operadores_gas = (
    df_gas.groupby("operador")["gas_natural_sin_nitrogeno_mmpcd"]
    .mean()
    .round(2)
    .sort_values(ascending=False)
    .head(10)
    .reset_index()
)

display(top10_operadores_gas)

# Guardar en Parquet
top10_operadores_gas.to_parquet(
    os.path.join(ruta_gold, "top10_operadores_gas.parquet"), index=False
)

# ================================
# Top operadores por producción total de ACEITE
# ================================
df_total_operador_aceite = (
    df_aceite.groupby("operador")["petroleo_mbd"]
    .sum()
    .round(2)
    .reset_index()
    .sort_values(by="petroleo_mbd", ascending=False)
)

print("🛢️ Top operadores por producción total de petróleo (Mbd):")
display(df_total_operador_aceite.head(10))

# Guardar en Parquet
df_total_operador_aceite.to_parquet(
    os.path.join(ruta_gold, "total_operador_aceite.parquet"), index=False
)

# ================================
# Top operadores por producción total de GAS
# ================================
df_total_operador_gas = (
    df_gas.groupby("operador")["gas_natural_sin_nitrogeno_mmpcd"]
    .sum()
    .round(2)
    .reset_index()
    .sort_values(by="gas_natural_sin_nitrogeno_mmpcd", ascending=False)
)

print("🔥 Top operadores por producción total de gas (MMpcd):")
display(df_total_operador_gas.head(10))

# Guardar en Parquet
df_total_operador_gas.to_parquet(
    os.path.join(ruta_gold, "total_operador_gas.parquet"), index=False
)

# ===============================
# 1. Operadores únicos - Aceite
# ===============================
operadores_aceite = sorted(df_aceite["operador"].unique())
df_operadores_aceite = pd.DataFrame(operadores_aceite, columns=["operador"])
print(f"Total de operadores únicos en aceite: {len(operadores_aceite)}")
display(df_operadores_aceite)


df_operadores_aceite.to_parquet(
    os.path.join(ruta_gold, "operadores_unicos_aceite.parquet"), index=False
)

# ===============================
# Campos únicos - Aceite
# ===============================
campos_aceite = sorted(df_aceite["campo"].unique())
df_campos_aceite = pd.DataFrame(campos_aceite, columns=["campo"])
print(f"Total de campos únicos en aceite: {len(campos_aceite)}")
display(df_campos_aceite)

df_campos_aceite.to_parquet(
    os.path.join(ruta_gold, "campos_unicos_aceite.parquet"), index=False
)


# ===============================
# Operadores únicos - Gas
# ===============================
operadores_gas = sorted(df_gas["operador"].unique())
df_operadores_gas = pd.DataFrame(operadores_gas, columns=["operador"])
print(f"Total de operadores únicos en gas: {len(operadores_gas)}")
display(df_operadores_gas)

df_operadores_gas.to_parquet(
    os.path.join(ruta_gold, "operadores_unicos_gas.parquet"), index=False
)

# ===============================
# Campos únicos - Gas
# ===============================
campos_gas = sorted(df_gas["campo"].unique())
df_campos_gas = pd.DataFrame(campos_gas, columns=["campo"])
print(f"Total de campos únicos en gas: {len(campos_gas)}")
display(df_campos_gas)

df_campos_gas.to_parquet(
    os.path.join(ruta_gold, "campos_unicos_gas.parquet"), index=False
)

# =======================
# 9. Guardado de datos limpios para ML
# =======================
display(df_aceite.head(2))
df_aceite.to_parquet(os.path.join(ruta_gold, "aceite_limpio.parquet"), index=False)
display(df_gas.head(2))
df_gas.to_parquet(os.path.join(ruta_gold, "gas_limpio.parquet"), index=False)


print("Años disponibles en df_aceite:", df_aceite["anio"].min(), "→", df_aceite["anio"].max())
print("Cantidad total de años distintos:", df_aceite["anio"].nunique())


print("Lista de años únicos:", sorted(df_aceite["anio"].unique()))
